backend/requirements.txt
audioread==3.0.0
Django==4.1.7
djangorestframework==3.14.0
django-cors-headers==3.14.0
librosa==0.10.0
moviepy==1.0.3
numpy==1.23.5
pytube==12.1.2
tensorflow
# tensorflow_macos
# soundfile==0.12.1
